ERR: On sgi, explict cast to double required for transcendtals.

+++ Code/BasicFilters/itkAtanImageFilter.h
@@ -37,7 +37,7 @@ namespace function {
     ~Atan() {};
     inline TOutput operator()( const TInput & A )
     {
-      return (TOutput)atan(A);
+      return (TOutput)atan((double) A);
     }
   };
